changed project structure, combined API Key related functions to APIKeys.js
