+++ src/lib/components/PackageCard.svelte
