+++ src/lib/data/userInitialsStore.js
