+++ src/routes/admin/+page.js
@@ -1,4 +1,4 @@
-import { initKeys, initUsers } from '../../stores/packstore';
+import { initKeys, initUsers } from '../../lib/stores/packstore';
 
 export async function load() {
     await initUsers();
